Revert "GraphicBufferSource: Use 64bit consumer usage"
This reverts commit b6caaf8dd20f4e79ccf74b6ac65ed2e52b70d4e5.

Reason for revert: b/295498401- AOSP phones not booting

Passing with revert: go/forrest-run/L01600000962534617

+++ media/libstagefright/omx/OmxGraphicBufferSource.cpp
@@ -85,7 +85,7 @@ status_t OmxGraphicBufferSource::configure(
         int32_t bufferCount,
         uint32_t frameWidth,
         uint32_t frameHeight,
-        uint64_t consumerUsage) {
+        uint32_t consumerUsage) {
     if (omxNode == NULL) {
         return BAD_VALUE;
     }
